Update for latest version of sveltekit
