+++ package.json
@@ -10,8 +10,8 @@
     "axios": "npm add axios"
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "1.0.0-next.64",
-    "@sveltejs/kit": "1.0.0-next.405",
+    "@sveltejs/adapter-auto": "next",
+    "@sveltejs/kit": "next",
     "autoprefixer": "^10.4.7",
     "postcss": "^8.4.14",
     "postcss-load-config": "^4.0.1",
@@ -23,6 +23,6 @@
   },
   "type": "module",
   "dependencies": {
-    "@storyblok/svelte": "^2.2.0"
+    "@storyblok/svelte": "2.3.5"
   }
 }
