+++ src/routes/home/+page.js
@@ -1,0 +1,5 @@
+import { redirect } from '@sveltejs/kit';
+
+export async function load() {
+    throw redirect(302, "/");
+}
